+++ Forge/src/main/java/com/christofmeg/brutalharvest/client/model/BrutalModelLayers.java
@@ -1,0 +1,12 @@
+package com.christofmeg.brutalharvest.client.model;
+
+import com.christofmeg.brutalharvest.CommonConstants;
+import net.minecraft.client.model.geom.ModelLayerLocation;
+import net.minecraft.resources.ResourceLocation;
+
+public class BrutalModelLayers {
+
+    public static final ModelLayerLocation THROWING_KNIFE = new ModelLayerLocation(
+            new ResourceLocation(CommonConstants.MOD_ID, "models/entity/throwing_knife"), "main");
+
+}