+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/entity/TomatoProjectileEntity.java
@@ -30,6 +30,7 @@ public class TomatoProjectileEntity extends ThrowableItemProjectile {
         super(EntityTypeRegistry.TOMATO_PROJECTILE.get(), livingEntity, pLevel);
     }
 
+    @Override
     protected @NotNull Item getDefaultItem() {
         return this.getItemRaw().getItem();
     }
@@ -42,6 +43,7 @@ public class TomatoProjectileEntity extends ThrowableItemProjectile {
         return new ItemParticleOption(ParticleTypes.ITEM, stack);
     }
 
+    @Override
     public void handleEntityEvent(byte p_37402_) {
         if (p_37402_ == 3) {
             ParticleOptions particleOptions = this.getParticle();
@@ -51,6 +53,7 @@ public class TomatoProjectileEntity extends ThrowableItemProjectile {
         }
     }
 
+    @Override
     protected void onHitEntity(@NotNull EntityHitResult entityHitResult) {
         super.onHitEntity(entityHitResult);
         Entity entity = entityHitResult.getEntity();
@@ -66,6 +69,7 @@ public class TomatoProjectileEntity extends ThrowableItemProjectile {
 
     }
 
+    @Override
     protected void onHit(@NotNull HitResult hitResult) {
         super.onHit(hitResult);
         this.playSound(SoundRegistry.TOMATO_SPLAT.get(), 0.5F, 0.8F);