+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/item/KnifeItem.java
@@ -1,12 +1,50 @@
 package com.christofmeg.brutalharvest.common.item;
 
+import com.christofmeg.brutalharvest.common.entity.ThrowingKnifeEntity;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.stats.Stats;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.projectile.AbstractArrow;
+import net.minecraft.world.entity.projectile.ThrownTrident;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.SwordItem;
 import net.minecraft.world.item.Tier;
+import net.minecraft.world.level.Level;
+import org.jetbrains.annotations.NotNull;
 
 public class KnifeItem extends SwordItem {
 
+    public float attackDamage;
+
     public KnifeItem(Tier tier, int attackDamage, Properties properties) {
         super(tier, attackDamage, -2, properties);
+        this.attackDamage = (float)attackDamage + tier.getAttackDamageBonus();
+    }
+
+    public @NotNull InteractionResultHolder<ItemStack> use(Level level, Player player, @NotNull InteractionHand interactionHand) {
+        ItemStack stack = player.getItemInHand(interactionHand);
+        level.playSound(null, player.getX(), player.getY(), player.getZ(), SoundEvents.SNOWBALL_THROW, SoundSource.NEUTRAL, 0.5F, 0.4F / (level.getRandom().nextFloat() * 0.4F + 0.8F));
+        if (!level.isClientSide) {
+            ThrowingKnifeEntity knifeEntity = new ThrowingKnifeEntity(level, player, stack, attackDamage);
+            knifeEntity.shootFromRotation(player, player.getXRot(), player.getYRot(), 0.0F, 1.5F, 1.0F);
+            level.addFreshEntity(knifeEntity);
+
+            ThrownTrident $$7 = new ThrownTrident(level, player, stack);
+            $$7.shootFromRotation(player, player.getXRot(), player.getYRot(), 0.0F, 2.5F, 1.0F);
+            if (player.getAbilities().instabuild) {
+                $$7.pickup = AbstractArrow.Pickup.CREATIVE_ONLY;
+            }
+        }
+
+        player.awardStat(Stats.ITEM_USED.get(this));
+        if (!player.getAbilities().instabuild) {
+            stack.shrink(1);
+        }
+
+        return InteractionResultHolder.sidedSuccess(stack, level.isClientSide());
     }
 
 }