+++ Forge/changelog.md
@@ -1,2 +1,4 @@
 1.6.0:
 * Added colored fabrics
+* Made knives throwable
+* Added "Throwing knives?" advancement